Added useListGroupSelect hook
